Move health page to be a layout
