+++ site/src/AppRouter.tsx
@@ -4,6 +4,7 @@ import AuditPage from "pages/AuditPage/AuditPage";
 import LoginPage from "pages/LoginPage/LoginPage";
 import { SetupPage } from "pages/SetupPage/SetupPage";
 import { TemplateLayout } from "pages/TemplatePage/TemplateLayout";
+import { HealthLayout } from "pages/HealthPage/HealthLayout";
 import TemplatesPage from "pages/TemplatesPage/TemplatesPage";
 import UsersPage from "pages/UsersPage/UsersPage";
 import WorkspacesPage from "pages/WorkspacesPage/WorkspacesPage";
@@ -197,9 +198,9 @@ const TemplateInsightsPage = lazy(
   () =>
     import("./pages/TemplatePage/TemplateInsightsPage/TemplateInsightsPage"),
 );
-const HealthPage = lazy(() => import("./pages/HealthPage/HealthPage"));
 const GroupsPage = lazy(() => import("./pages/GroupsPage/GroupsPage"));
 const IconsPage = lazy(() => import("./pages/IconsPage/IconsPage"));
+const AccessURLPage = lazy(() => import("./pages/HealthPage/AccessURLPage"));
 
 export const AppRouter: FC = () => {
   return (
@@ -213,8 +214,6 @@ export const AppRouter: FC = () => {
           <Route element={<RequireAuth />}>
             <Route element={<DashboardLayout />}>
               <Route index element={<Navigate to="/workspaces" replace />} />
-
-              <Route path="/health" element={<HealthPage />} />
 
               <Route
                 path="/external-auth/:provider"
@@ -339,6 +338,11 @@ export const AppRouter: FC = () => {
                     />
                   </Route>
                 </Route>
+              </Route>
+
+              <Route path="/health" element={<HealthLayout />}>
+                <Route index element={<Navigate to="access-url" />} />
+                <Route path="access-url" element={<AccessURLPage />} />
               </Route>
             </Route>
