+++ site/src/pages/HealthPage/AccessURLPage.tsx
@@ -1,0 +1,5 @@
+export const AccessURLPage = () => {
+  return <div>Access URL</div>;
+};
+
+export default AccessURLPage;
